requirements.txt
streamlit>=1.43.0
pdfplumber>=0.11.4
pandas>=2.2.3
numpy>=2.2.0
plotly>=5.24.0
anthropic>=0.89.0
scikit-learn>=1.6.1
ofxparse>=0.21
openpyxl>=3.1.5
python-dotenv>=1.0.1
pyyaml>=6.0.2
pydantic>=2.10.6
click>=8.1.8
rich>=13.9.4
